+++ relay-general/src/protocol/contexts/browser.rs
@@ -1,0 +1,31 @@
+use once_cell::sync::OnceCell;
+use regex::Regex;
+use serde::{Serialize, Serializer};
+
+use crate::processor::ProcessValue;
+use crate::protocol::LenientString;
+use crate::types::{
+    Annotated, Empty, Error, FromValue, IntoValue, Object, SkipSerialization, Value,
+};
+
+/// Web browser information.
+#[derive(Clone, Debug, Default, PartialEq, Empty, FromValue, IntoValue, ProcessValue)]
+#[cfg_attr(feature = "jsonschema", derive(JsonSchema))]
+pub struct BrowserContext {
+    /// Display name of the browser application.
+    pub name: Annotated<String>,
+
+    /// Version string of the browser.
+    pub version: Annotated<String>,
+
+    /// Additional arbitrary fields for forwards compatibility.
+    #[metastructure(additional_properties, retain = "true", pii = "maybe")]
+    pub other: Object<Value>,
+}
+
+impl BrowserContext {
+    /// The key under which a browser context is generally stored (in `Contexts`)
+    pub fn default_key() -> &'static str {
+        "browser"
+    }
+}
